delete endpoint

+++ src/main/java/com/example/TaskManagement/repository/UsersRepository.java
@@ -3,7 +3,7 @@ package com.example.TaskManagement.repository;
 import com.example.TaskManagement.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepository extends JpaRepository<User, Long> {
+public interface UsersRepository extends JpaRepository<User, Integer> {
 
 
 }
